Add files via upload
This project is an 'in  class assignment' given to me (and my classmates - separately) to solve and answer in order to demonstrate our individual capabilities in R and our knowledge/understanding of  advance econometrics and advanced analytical tools.

I must reiterate I did this on my own. This is my individual copy of a class assigned activity or project.
I hope you like it.